+++ tests/unit/test_ska_dlm_client.py
@@ -25,4 +25,4 @@ def test_content(version: str) -> None:
 
     :param version: the version fixture
     """
-    assert version == "1.0.0"
+    assert version == "1.1.0"